feature(jwk): add rcrypto::Key
This represents a fully parsed key. It is an enum containing
fully-parsed key material with algorithm types and public/private types
chosen at runtime rather than at compile time. Thinking of it a
different way, a Key is an abstract representation of a cryptographic
key with the same invariants as a JWK.

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>